backend/requirements.txt
asgiref
celery
cloudinary
Django
django-cloudinary-storage
django-cors-headers
django-environ
django-redis
djangorestframework
djangorestframework-simplejwt
kombu
psycopg2-binary
python-dotenv
redis
requests
vine